+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 * Improve header actions performance (icons are pure components).
 * When selecting subjects, back button clears selection.
+* When importing, archived subjects are considered.
 
 
 ## 1.6.0 - 2019-06-01